+++ data/src/main/java/com/youhajun/data/models/dto/store/PurchaseInAppItem.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class PurchaseInAppItem(
     @SerializedName("idx")
-    private val _idx: Int?,
+    private val _idx: Long?,
 
     @SerializedName("productId")
     private val _productId: String?,
@@ -21,7 +21,7 @@ data class PurchaseInAppItem(
     @SerializedName("itemCount")
     private val _itemCount: Int?,
 ) : PurchaseItem {
-    override val idx: Int
+    override val idx: Long
         get() = _idx ?: -1
     override val productId: String
         get() = _productId ?: ""
