+++ ui/src/main/java/com/youhajun/ui/models/sideEffects/SelectRoomSideEffect.kt
@@ -4,6 +4,6 @@ sealed class SelectRoomSideEffect {
     data class Toast(val text: String) : SelectRoomSideEffect()
     sealed class Navigation: SelectRoomSideEffect() {
         object NavigateUp : Navigation()
-        data class NavigateToLiveRoom(val idx:Int) : Navigation()
+        data class NavigateToLiveRoom(val idx:Long) : Navigation()
     }
 }
